pyproject.toml
[project]
name = "bithumb-event"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "cozy-hn", email = "jhk2721@gmail.com"},
]
dependencies = ["ccxt>=4.5.19", "python-dotenv>=1.2.1"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
